refactor: replace `Wrap` struct with `Pair` and update aggregate actions
- Renamed `Wrap` to `Pair` for better clarity.
- Refactored helper functions to `numberToPair` and `messageToPair`.
- Updated tests to reflect the new struct and helper function names.
